Save & Retrieve data in Electric Meter Details Report complete.

+++ app/src/main/java/com/example/mpcb/network/request/ReportRequest.kt
@@ -335,7 +335,7 @@ class RoutineReport {
     var electrictMeterProvided: Int? = null
     @SerializedName("electrict_meter_reading")
     @Expose
-    var electrictMeterReading: Int? = null
+    var electrictMeterReading: Double? = null
 
 
     @SerializedName("date_of_collection_indus")
